# api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: imajin events
  version: "1.0.0"
  description: >
    Event creation, ticketing, ticket tiers, queue management, hold system, .fair manifests, checkout, and payment webhooks.
    Served by the app itself at /api/spec. Authenticate with a scoped app token (`Authorization: Bearer`, audience `events`,
    scopes `events:read` for GET/HEAD/OPTIONS and `events:write` otherwise) or, during migration, the legacy session cookie.
servers:
  - url: https://events.imajin.ai
    description: production
  - url: https://dev-events.imajin.ai
    description: dev
  - url: http://localhost:3000
    description: local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: imajin_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Scoped app token minted for audience `events` (events:read for safe methods, events:write otherwise)"
    webhookSecret:
      type: http
      scheme: bearer
      description: "Bearer WEBHOOK_SECRET — internal payment webhook from pay service"
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Not authorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Event:
      type: object
      properties:
        id: { type: string }
        did: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [draft, published, paused, cancelled, completed] }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, nullable: true }
        timezone: { type: string, nullable: true }
        isVirtual: { type: boolean }
        virtualUrl: { type: string, nullable: true }
        venue: { type: string, nullable: true }
        address: { type: string, nullable: true }
        city: { type: string, nullable: true }
        country: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        fairManifest: { type: object }
        createdAt: { type: string, format: date-time }
    TicketTier:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        currency: { type: string }
        quantity: { type: integer, nullable: true }
        sold: { type: integer }
        available: { type: integer, nullable: true }
        perks: { type: array, items: { type: string } }
        requiresRegistration: { type: boolean }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
paths:
...
  /api/events:
    get:
      tags: [events]
      operationId: listEvents
      summary: List events
      parameters:
        - name: status
          in: query
          schema: { type: string, default: published }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: courseSlug
          in: query
          schema: { type: string }
        - name: upcoming
          in: query
          schema: { type: boolean }
          description: "Filter to events that haven't started yet"
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/Event" }
    post:
      tags: [events]
      operationId: createEvent
      summary: Create an event — requires hard DID
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, startsAt]
              properties:
                title: { type: string }
                description: { type: string }
                startsAt: { type: string, format: date-time }
                endsAt: { type: string, format: date-time }
                timezone: { type: string }
                isVirtual: { type: boolean }
                virtualUrl: { type: string }
                venue: { type: string }
                address: { type: string }
                city: { type: string }
                country: { type: string }
                imageUrl: { type: string }
                imageAssetId: { type: string }
                tags: { type: array, items: { type: string } }
                courseSlug: { type: string }
                emtEmail: { type: string }
                tickets:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      description: { type: string }
                      price: { type: number }
                      currency: { type: string }
                      quantity: { type: integer }
                      perks: { type: array, items: { type: string } }
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  event: { $ref: "#/components/schemas/Event" }
                  ticketTypes:
                    type: array
                    items: { $ref: "#/components/schemas/TicketTier" }
                  eventKeypair:
                    type: object
                    properties:
                      publicKey: { type: string }
                      privateKey: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Requires hard DID, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /api/migrate-tickets:
    post:
      tags: [events]
      operationId: migrateTickets
      summary: "Migrate soft-DID tickets to a hard DID after email verification (app token only)"
      description: "Requires a scoped app token carrying events:write. Always fails closed."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email, hardDid], properties: { email: { type: string }, hardDid: { type: string } } }
      responses:
        "200": { description: Tickets migrated (idempotent), content: { application/json: { schema: { type: object } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { description: Migration failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...